write manifest when export and share project
