+++ translationRecorder/app/src/main/java/org/wycliffeassociates/translationrecorder/FilesPage/Export/AppExport.kt
@@ -4,6 +4,8 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import androidx.core.content.FileProvider
+import org.wycliffeassociates.translationrecorder.database.IProjectDatabaseHelper
+import org.wycliffeassociates.translationrecorder.persistance.AssetsProvider
 import org.wycliffeassociates.translationrecorder.persistance.IDirectoryProvider
 import org.wycliffeassociates.translationrecorder.project.Project
 import java.io.File
@@ -13,8 +15,10 @@ import java.io.File
  */
 class AppExport(
     project: Project,
-    directoryProvider: IDirectoryProvider
-) : Export(project, directoryProvider) {
+    directoryProvider: IDirectoryProvider,
+    db: IProjectDatabaseHelper,
+    assetsProvider: AssetsProvider
+) : Export(project, directoryProvider, db, assetsProvider) {
 
     override fun handleUserInput() {
         exportZipApplications(zipFile!!)
